refactor: add sites module
