+++ main.lua
@@ -21,6 +21,9 @@ require("modules/update")
 require("apis/dandanplay")
 require('apis/extra')
 
+require("sites/bilibili")
+require("sites/bahamut")
+
 DANMAKU_PATH = os.getenv("TEMP") or "/tmp/"
 HISTORY_PATH = mp.command_native({"expand-path", options.history_path})
 PID = utils.getpid()
@@ -110,51 +113,6 @@ function show_loaded(init)
         show_message("弹幕加载成功，共计" .. #COMMENTS .. "条弹幕", 3)
     end
     mp.set_property_native(DANMAKU_COUNT, #COMMENTS)
-end
-
-local function get_cid()
-    local cid, danmaku_id = nil, nil
-    local tracks = mp.get_property_native("track-list")
-    for _, track in ipairs(tracks) do
-        if track["lang"] == "danmaku" then
-            cid = track["external-filename"]:match("/(%d-)%.xml$")
-            danmaku_id = track["id"]
-            break
-        end
-    end
-    return cid, danmaku_id
-end
-
-local function extract_between_colons(input_string)
-    local start_index = 0
-    local end_index = 0
-    local count = 0
-    for i = 1, #input_string do
-        if input_string:sub(i, i) == ":" then
-            count = count + 1
-            if count == 2 then
-                start_index = i
-            elseif count == 3 then
-                end_index = i
-                break
-            end
-        end
-    end
-    if start_index > 0 and end_index > 0 then
-        return input_string:sub(start_index + 1, end_index - 1)
-    else
-        return nil
-    end
-end
-
-local function get_type_from_position(position)
-    if position == 0 then
-        return 1
-    end
-    if position == 1 then
-        return 4
-    end
-    return 5
 end
 
 -- 获取指定时间的延迟
@@ -558,256 +516,6 @@ function load_danmaku(from_menu, no_osd)
     if not ENABLED then return end
     convert_danmaku_to_ass_events()
     render_danmaku(from_menu, no_osd)
-end
-
-local function get_bilibili_id_and_page(path)
-    local bvid, aid, page = nil, nil, nil
-    if not bvid then
-        bvid = path:match("/video/(BV[%w]+)")
-            or path:match("[?&]bvid=(BV[%w]+)")
-    end
-    if not aid then
-        aid = path:match("/video/av([%d]+)")
-    end
-    if not page then
-        page = tonumber(path:match("[?&]p=(%d+)"))
-    end
-    return bvid, aid, page or 1
-end
-
-local function get_bilibili_pagelist_args(bvid, aid)
-    local url
-    if bvid ~= nil then
-        url = "https://api.bilibili.com/x/player/pagelist?bvid=" .. bvid
-    else
-        url = "https://api.bilibili.com/x/player/pagelist?aid=" .. aid
-    end
-    local arg = {
-        "curl",
-        "-L",
-        "-s",
-        "--compressed",
-        "--user-agent",
-        "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/124.0.0.0 Safari/537.36 Edg/124.0.0.0",
-        "-H",
-        "Referer: https://www.bilibili.com/video/" .. (bvid or ("av" .. aid)),
-        url
-    }
-    if options.cookie_file and options.cookie_file ~= "" then
-        table.insert(arg, '-b')
-        table.insert(arg, mp.command_native({"expand-path", options.cookie_file}))
-    end
-
-    if options.proxy ~= "" then
-        table.insert(arg, '-x')
-        table.insert(arg, options.proxy)
-    end
-
-    return arg
-end
-
-local function resolve_bilibili_cid(path, callback)
-    local bvid, aid, page = get_bilibili_id_and_page(path)
-    if not bvid and not aid then
-        callback(nil)
-        return
-    end
-
-    call_cmd_async(get_bilibili_pagelist_args(bvid, aid), function(error, json)
-        if error then
-            msg.warn("Failed to request bilibili pagelist: " .. tostring(error))
-            callback(nil)
-            return
-        end
-
-        local data = utils.parse_json(json)
-        local pages = data and data["data"]
-        if type(pages) ~= "table" then
-            callback(nil)
-            return
-        end
-
-        local page_info = pages[page] or pages[1]
-        local cid = page_info and page_info["cid"]
-        callback(cid and tostring(cid) or nil)
-    end)
-end
-
-local function download_bilibili_danmaku(path, cid, from_menu)
-    local url = "https://comment.bilibili.com/" .. cid .. ".xml"
-    local temp_file = "danmaku-" .. PID .. DANMAKU.count .. ".xml"
-    local danmaku_xml = utils.join_path(DANMAKU_PATH, temp_file)
-    DANMAKU.count = DANMAKU.count + 1
-    local arg = {
-        "curl",
-        "-L",
-        "-s",
-        "--compressed",
-        "--user-agent",
-        "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/124.0.0.0 Safari/537.36 Edg/124.0.0.0",
-        "--output",
-        danmaku_xml,
-        url,
-    }
-
-    if options.cookie_file and options.cookie_file ~= "" then
-        table.insert(arg, '-b')
-        table.insert(arg, mp.command_native({"expand-path", options.cookie_file}))
-    end
-
-    if options.proxy ~= "" then
-        table.insert(arg, '-x')
-        table.insert(arg, options.proxy)
-    end
-
-    call_cmd_async(arg, function(error)
-        if error then
-            show_message("HTTP request failed, see console for details", 5)
-            msg.error(error)
-            return
-        end
-        if file_exists(danmaku_xml) then
-            save_danmaku_downloaded(path, danmaku_xml)
-            load_danmaku(from_menu == nil and true or from_menu)
-        end
-    end)
-end
-
--- 为 bilibli 网站的视频播放加载弹幕
-function load_danmaku_for_bilibili(path)
-    local cid, danmaku_id = get_cid()
-    if danmaku_id ~= nil then
-        mp.commandv('sub-remove', danmaku_id)
-    end
-
-    if cid == nil then
-        cid = mp.get_opt('cid')
-        if not cid then
-            local patterns = {
-                "bilivideo%.c[nom]+.*/resource/(%d+)%D+.*",
-                "bilivideo%.c[nom]+.*/(%d+)-%d+-%d+%..*%?",
-            }
-            local urls = {
-                path,
-                mp.get_property("stream-open-filename", ''),
-            }
-
-            for _, pattern in ipairs(patterns) do
-                for _, url in ipairs(urls) do
-                    if url:find(pattern) then
-                        cid = url:match(pattern)
-                        break
-                    end
-                end
-            end
-        end
-    end
-    if cid == nil then
-        resolve_bilibili_cid(path, function(resolved_cid)
-            if resolved_cid then
-                download_bilibili_danmaku(path, resolved_cid, true)
-            else
-                msg.error("获取哔哩哔哩视频cid失败")
-            end
-        end)
-        return
-    end
-    if cid ~= nil then
-        download_bilibili_danmaku(path, cid, true)
-    end
-end
-
--- 为 bahamut 网站的视频播放加载弹幕
-function load_danmaku_for_bahamut(path)
-    local path = path:gsub('%%(%x%x)', hex_to_char)
-    local sn = extract_between_colons(path)
-    if sn == nil then
-        return
-    end
-    local url = "https://ani.gamer.com.tw/ajax/danmuGet.php"
-    local temp_file = "bahamut-" .. PID .. ".json"
-    local danmaku_json = utils.join_path(DANMAKU_PATH, temp_file)
-    local arg = {
-        "curl",
-        "-X",
-        "POST",
-        "-d",
-        "sn=" .. sn,
-        "-L",
-        "-s",
-        "--user-agent",
-        "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/85.0.4183.83 Safari/537.36",
-        "--header",
-        "Origin: https://ani.gamer.com.tw",
-        "--header",
-        "Content-Type: application/x-www-form-urlencoded;charset=utf-8",
-        "--header",
-        "Accept: application/json",
-        "--header",
-        "Authority: ani.gamer.com.tw",
-        "--output",
-        danmaku_json,
-        url,
-    }
-
-    if options.proxy ~= "" then
-        table.insert(arg, '-x')
-        table.insert(arg, options.proxy)
-    end
-
-    if options.cookie_file and options.cookie_file ~= "" then
-        table.insert(arg, '-b')
-        table.insert(arg, mp.command_native({"expand-path", options.cookie_file}))
-    end
-
-    call_cmd_async(arg, function(error)
-        if error then
-            show_message("HTTP 请求失败，打开控制台查看详情", 5)
-            msg.error(error)
-            return
-        end
-        if not file_exists(danmaku_json) then
-            return
-        end
-
-        local comments_json = read_file(danmaku_json)
-        os.remove(danmaku_json)
-        local comments = utils.parse_json(comments_json)
-        if not comments then
-            return
-        end
-
-        local output_table = {}
-        for _, comment in ipairs(comments) do
-            local color = hex_to_int_color(comment["color"])
-            local mode = get_type_from_position(comment["position"])
-            local time = tonumber(comment["time"]) / 10
-            local c_param = string.format("%s,%s,%s,25,,,", time, color, mode)
-            table.insert(output_table, {
-                c = c_param,
-                m = comment["text"]
-            })
-        end
-
-        local final_json_str = utils.format_json(output_table)
-
-        temp_file = "danmaku-" .. PID .. DANMAKU.count .. ".json"
-        local json_filename = utils.join_path(DANMAKU_PATH, temp_file)
-        DANMAKU.count = DANMAKU.count + 1
-
-        local json_file = io.open(json_filename, "w")
-        if json_file then
-            json_file:write(final_json_str)
-            json_file:close()
-        end
-
-        if file_exists(json_filename) then
-            save_danmaku_downloaded(
-                "https://ani.gamer.com.tw/animeVideo.php?sn=" .. sn,
-                json_filename)
-            load_danmaku(true)
-        end
-    end)
 end
 
 function load_danmaku_for_url(path)
